[Rename] 폴더명 수정
