+++ src/main/java/com/example/shopping/exception/item/ItemNotFoundException.java
@@ -1,4 +1,4 @@
-package com.example.shopping.exception;
+package com.example.shopping.exception.item;
 
 public class ItemNotFoundException extends ItemException{
     public ItemNotFoundException(String message) {
